Rework application for the official Phoenix v1.3.0 release
The application was actually started from scratch, and this commit represents a diff between what would of resulted from following the blog post using Phoenix v1.3.0-rc2 and Phoenix v1.3.0. The Phoenix v1.3.0-rc2 version of the application is tagged as phoenix_v1.3.0-rc2 and is the parent of this commit.